better arrange the tests
